+++ Editor/Playa/Renderer/RealTimeCalculatorFakeRenderer/RealTimeCalculatorRendererUIToolkit.cs
@@ -94,6 +94,8 @@ namespace SaintsField.Editor.Playa.Renderer.RealTimeCalculatorFakeRenderer
 
             root.RegisterCallback<AttachToPanelEvent>(_ => UIToolkitUtils.LoopCheckOutOfScoopFoldout(root));
 
+            string targetId = $"{FieldWithInfo.Targets[0].GetHashCode()}.{methodInfo.Name}";
+
             if (hasParameters)
             {
                 _ussClassSaintsFieldEditingDisabledHide ??= Util.LoadResource<StyleSheet>("UIToolkit/ClassSaintsFieldEditingDisabledHide.uss");
@@ -146,7 +148,8 @@ namespace SaintsField.Editor.Playa.Renderer.RealTimeCalculatorFakeRenderer
                             InAnyHorizontalLayout,
                             attributes,
                             FieldWithInfo.Targets,
-                            this
+                            this,
+                            $"{targetId}.{parameterInfo.Name}"
                         ).result;
                         // ReSharper disable once InvertIf
                         if (r != null)
@@ -300,6 +303,7 @@ namespace SaintsField.Editor.Playa.Renderer.RealTimeCalculatorFakeRenderer
                 }
 
                 Type type = RuntimeUtil.IsNull(value) ? methodInfo.ReturnType : value.GetType();
+                string targetId = $"{FieldWithInfo.Targets[0].GetHashCode()}.{methodInfo.Name}.{type.FullName}";
                 (VisualElement result, bool isNestedField) = UIToolkitEdit.UIToolkitValueEdit(
                     fieldElementOrNull,
                     NoLabel? null: GetName(FieldWithInfo),
@@ -311,7 +315,8 @@ namespace SaintsField.Editor.Playa.Renderer.RealTimeCalculatorFakeRenderer
                     InAnyHorizontalLayout,
                     ReflectCache.GetCustomAttributes(FieldWithInfo.MethodInfo),
                     FieldWithInfo.Targets,
-                    this);
+                    this,
+                    targetId);
                 // Debug.Log($"fill value {value} get {result}");
                 if(result!=null)
                 {